Simplify project structure
